requirements.txt
# 核心依赖
numpy>=1.20.0
pandas>=1.3.0
polars>=0.8.0
scipy>=1.7.0
scikit-learn>=1.0.0

# DolphinDB相关
dolphindb>=1.30.17.1

# LLM相关
openai>=1.0.0
langchain>=0.0.267
transformers>=4.30.0
sentence-transformers>=2.2.0
pypdf2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10

# 回测相关
pyfolio>=0.9.2
empyrical>=0.5.5

# 数据处理
pyarrow>=8.0.0
fastparquet>=0.8.0
openpyxl>=3.0.9

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
dash>=2.6.0
dash-bootstrap-components>=1.2.0

# API和Web服务
fastapi>=0.89.0
uvicorn>=0.20.0
pydantic>=1.10.0
jinja2>=3.0.0

# 工具类
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=0.20.0
loguru>=0.6.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0

# 开发工具
black>=22.10.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.991
